// enums/queryKeys.ts
export enum QueryKeys {
  Categories = 'categories',
  RecipesByTime = 'recipesByTime',
  RecipePhoto = 'recipePhoto',
  RecipesByRating = 'recipesByRating',
  Recipes = 'recipes',
  FilteredRecipes = 'filteredRecipes',
  RecipeRating = 'recipeRating',
  SavedRecipesIds = 'savedRecipesIds',
  SavedRecipes = 'savedRecipes',
  UserProfilePicture = 'userProfilePicture'
}